Add GoalPlatform, NoclipEnabler, and SpawnPlatform templates (#11)
### Changes
As discussed on Tuesday, we introduce 3 entities with associated tags.

1) `GoalPlatform `-> when this element is reached, we win the level
2) `NoclipEnabler `-> the player can activate noclip mode only when it
is here, and must return here when noclip mode time is over
3) `SpawnPlatform `-> the player spawns from here

### Remarks
- The purpose of this PR is to introduce the key tags and the structure.
The materials/shape/properties of the objects will be discussed later on
- We need to decide where to write the logic of the interactions above.
For instance: who decides where the player spawns? The player controller
or the level manager? Which script decides when the noclip mode can be
activated and deactivated? Etc...
- I think that, after this PR, we have all the elements to write all the
dynamics of the demo level

Co-authored-by: spig95 <nicola.ischia@schindler.com>