Refactored directories to include player management, updated prefabs and GoalPlatform
In addition to the save feature, I've deleted the FinishLevel script which was redundant and I've added yet another collision check on the player to optimize code readibility.

On top of that I've refactored directories to better understand where player controls reside.

I've then updated the GoalPlatform and SpawnPlatform prefab to include the ProgressBox.