Save feature added and various tweaks (#38)
### Finally, saves!
This is a _small_ PR which implements a save/resume feature in noclip
and various tweaks. Although pretty basic, I think that for now it's
working as intended and it could be stylistically improved by adding
some sort of canvas which enables certain icons when saving the game.

Here's a brief recap:
- Added a save feature in `RealityPlayerCollisions.cs`.
- Refactored some directories, now all player centered scripts will
reside in the **PlayerManagement** directory, while gameplay-focused
aspects will reside in the **NoclipRealityManagement** directory. Let me
know if this sounds right.
- Updated `GoalPlatform` and `SpawnPlatform` prefabs. 
- As a consequence, `FinishLevel.cs` has been deleted since it was
redundant to have a collision script with a manager for player's
collisions. A `ProgressBox` with its respective tag has been added to
the `SpawnPlatform` prefab.
- Polished some comments for readability purposes.

What's next? Well, as I've mentioned above, some sort of canvas which
will dynamically change based on various events. For example, if a save
occurs, it'll be shown with an icon on the bottom left part of said
canvas.