Added audio, sensitivity and inverted axis methods. Ditched brightness, for now (#49)
### It just works
Unironically, it does. Here are the changes I've done to implement these
methods:
- **Added working audio settings**. They'll work throughout the levels
and are resilient to the saving mechanic.
- **Added working sensitivity and axis inversion methods**! It's just a
multiplication, but it does the trick.
- As a result, I've updated `MouseLook.cs` to implement a sensitivity
and axis inversion method.
- I've also done some more polishing in the comments, particularly in
`MouseLook.cs`.
- MenuController and LoadPrefs have been updated to accomodate these
changes, and its summary has been revisited.
- Oh, uhh, the menu has a new save icon. It's really cool. I would dare
to say it's the best piece of art I've ever seen around since the dawn
of men.

### It just doesn't work
- Brightness. The most fearsome enemy a man can fight. I have nightmares
about it. From now on I'll live in fear. Please help.
- Jokes aside, brightness has been temporarily ditched. There's an issue
regarding this discussion, and if anyone wants to give his feedback,
well... it would be appreciated.

Now, it would be a shame if I had a plethora of conflicts while
merging...
.................................
.................................
.................................
................................................. **_right_**?

Co-authored-by: Nicola Ischia <36990714+spig95@users.noreply.github.com>