Add a new noclip object type and the related controller script (#48)
### Changes
- Move the scripts that control the noclip objects in the
`NoclipObjectControllers` folder
- Add a parent abstract class `BaseNoclipObjectController.cs` and make
the `NoclipObjController.cs` a child of this class.
- Create an intangible object prefab. It has the NoclipObject tag and
the brand new script `NoclipIntangibleController.cs`, which is also a
child of the base controller class for Noclip objects
- Minor changes to scene_3 (remove stairs and add the new intangible
object prefab)

### Remarks
Having a base class `BaseNoclipObjectController.cs` is particularly
useful: the child class highlights the difference in the no clip object
types, while the abstract one highlight the interface to the rest of the
world.

Moreover, in `NoclipManager`, we can make everything work with minor
modifications. As you can see from the changes, we just take all the
objects with the NoclipObject tag and get their components of type
`BaseNoclipObjectController` (we exploit inheritance: the childs have
the same type of the father). The NoclipManager does not need to know
the different Noclip object types!

This makes it really easy to add other noclip behaviours by simply
creating a child noclip object controller class and attaching it to a
given object with the `NoclipObject` tag