Audio clips implemented through scriptable object (#143)
### At last, audio
- Reworked the audio system. Audio sources have been distributed where
needed.
- A new scriptable object, `DefaultAudioTracks.cs`, has been added. It
contains many tracks used for various purposes, such as player feedback
and level soundtracks.
- The noclip enabler zone, platform and noclip mode now have their
distinct sounds.
- Footsteps have been converted to the new audio system, and they come
with an _interesting easter egg_.
- A few placeholder clips have been added.
- A neat trick which lets you pick a custom soundtrack for a given level
has been added. If there is none, a default tune will play.
- And, finally, the effects channel now serves its purpose.

### Remarks
Adding clips, tracks and other effects has now become much easier and
efficient.
There still needs to be an actual effect system where things change when
certain things happen, for example the infamous sound muffling when in
noclip mode. More on that in another PR.

### Improvement areas
As I've already said, effects are still **not** implemented. 
The audio mixer is cool, and provides much more tuning capabilities, but
I feel that some channels are not really useful. With that in mind,
channels will have to be reworked a little to trim off unnecessary
parts.

Co-authored-by: spig95 <nicolaischia95@gmail.com>