Always return to realitycamera rotation after noclip with animation (#149)
Add `IsBackToBody()` function which checks also the rotation, not only
the distance in the space.

In this way, if we enable noclip and we just change the rotation
(without changing position), we also have animation when we disable
noclip. The animation resets the reality orientation to the one we had
before enabling noclip.