Post Processing effects now toned down for noclip mode + menu fixes (#166)
### Noclip: Post Processing Evolved
Hi, you beautiful, mesmerizing, sensational souls. Here's a brief recap
of this PR, which fixes a few things here and there and adds the (basic)
option to have a noclip effect and a reality effect.
- I've disabled bloom on noclip mode, for now. We'll see before
submitting the final prototype.
- Fixed some bugs on the menu: as @TizioMaurizio noticed, it was
bothersome to return to the start from the outside part of the menu, so
I've added the option to return from whenever you are.
- Controls do not appear on the main menu, and you'll have to manually
select them. This was a design choice, and I'm confident it does look
better and not overcrowded this way.

### Y
- Invert Y doesn't work. I can't understand why. I've modified the
`updateInvertY` function in order to accomodate events, but as
@TizioMaurizio mentioned, `mouselook.cs` doesn't update the Y, therefore
the cause isn't clear at all at this time. I'll have a look with
@stefanofossati tomorrow... I'll fight for this feature!

Co-authored-by: TizioMaurizio <maurizio.vetere@mail.polimi.it>
Co-authored-by: stefanofossati <10569836@mail.polimi.it>