Tutorial and narrative dialogues showing up + fade-in/out now working! (#206)
Long time no see. 

At last, the tutorial dialogue has been overhauled. It's not final, of
course, but it'll be soon. The foundation is there and it's working!

Here's a brief recap:
- Added the final Area 00 puzzle. (Note: it'll be divided into two
sections)
- Tutorials and dialogues added! Everything is a placeholder, but the
foundation is working as intended. When stepping into a triggered area,
a dialogue will appear. Said dialogue could or could not imply a
movement reduction depending on design choices. Either way, it can be
skipped by pressing the dedicated skip button (for now, it's backspace.
It's subject to change).
- Added checks for tutorial in the event manager. This makes passing
parameters easy.
- Fade-in/out has been added. Coroutines management has been tricky, but
it paid off and it looks nice. The only downside is the canvas: it isn't
working as intended, therefore the dialogue pops up differently based on
your resolution / aspect-ratio.
- Just like a typewriter, dialogue pops with a sequence which lasts as
long as the text goes. Keep reading!
- Code optimizations.

What's left?
- Crosshair. It still doesn't count the time visually (colors).
- Audio cues for dialogue. 
- UI fixes.

Co-authored-by: TizioMaurizio <maurizio.vetere@mail.polimi.it>
Co-authored-by: Alessio De Nicolò <alessio1.denicolo@mail.polimi.it>
Co-authored-by: stefanofossati <10569836@mail.polimi.it>
Co-authored-by: spig95 <nicolaischia95@gmail.com>