Loading screen implementation (#209)
### Welcome, Narrative Section
I know what you're thinking: this is a PR made for loading screens. 
You couldn't be more *wrong*. 

This is actually a fake loading screen. We made the timer go along with
a pre-determined time to give the illusion of a loading screen, where it
actually serves as a narration sequence.

### Recap
- Added loading screen.
- Added loading animation.
- Whenever loading is finished, you can press any button to continue.
- Added a scriptable object to better control which text to be shown in
a given level.

### What's left
- Dialogue: we'll have to consult with @AleDeNic in order to finalize
this aspect.
- Clunky fade-in: we'll have to investigate further ways to fade text.
- Typewriter effect. A subtle effect will have to be added to better
reflect tutorial prompts.

Co-authored-by: stefanofossati <10569836@mail.polimi.it>