Simplify TimeConstraints and fix all the bugs (hopefully!) (#228)
There were many bugs in my previous implementation of `TimeConstraints`
since I did not properly update it after the changes to the noclip
manager.

Now, I have decoupled it from the noclip manager and I have added an
event (**ResumeTimeConstraintsTimer**) and renamed
StartTimeConstraintsTimer into **RestartTimeConstraintsTimer**. In
addition, **PauseTimeConstraintsTimer** can be used to pause the time
constraints. Before there was not such a function.

Maybe there are too many events there, and for sure this is not the most
elegant solution. But I hope it makes it easier to handle and removes
many bugs.. please let me know!