Update time limits according to difficulty at each reset (#302)
Previously, the time for a puzzle was computed only when the puzzle was
unlocked, and not after each reset. Therefore, if the player was in
difficult mode and they did not manage to finish the puzzle, changing
the difficulty would not give them more time after a re-trial of the
puzzle.

Now, if the user changes difficulty, dies, and respawns, they will have
the time computed according to the new difficult level.