Cap the max time for which the clock effect can run to 4 seconds (#305)
This is a really minor change, I don't even know if you want it as a
feature or not. For me, it was annoying to have a really long clock
effect when a puzzle has a very long max time to complete. This PR cuts
the maximum "clock effect time" to 4 seconds (as a default)